refactor: Extract weights command to separate module and simplify numpy import
- Move weights command and related helper functions to new affine/weights.py module
- Simplify numpy import in storage.py (remove conditional import as numpy is a required dependency)
- Improve code organization and maintainability by separating CLI concerns

Changes:
- Created affine/weights.py with weights command logic and helper functions
- Updated affine/cli.py to import and register weights command from new module
- Removed try/except numpy import in affine/storage.py (numpy v2.3.4 is installed)
- Removed HAS_NUMPY check in _convert_numpy_types function
